refactor(dashboard): :art: Improved file structure for dashboard
