+++ app/modules/dashboard/components/atoms/avatar/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import * as AvatarPrimitive from "@radix-ui/react-avatar";
-import { cn } from "../../lib/utils";
+import { cn } from "../../../lib/utils";
 
 const Avatar = React.forwardRef<
   React.ElementRef<typeof AvatarPrimitive.Root>,
