+++ app/modules/dashboard/components/atoms/progressCircle/index.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as ProgressPrimitive from "@radix-ui/react-progress";
 
-import { cn } from "../../lib/utils"; // Usar la ruta relativa
+import { cn } from "../../../lib/utils"; // Usar la ruta relativa
 
 const Progress = React.forwardRef<
   React.ElementRef<typeof ProgressPrimitive.Root>,
